chore(tile): rename internal context file
